clip_vqa_english/requirements.txt
matplotlib
numpy
pandas
Pillow
scikit_learn
torch
torchvision
tqdm
transformers
wandb